+++ src/MCDRpost/mcdrpost/manager/post_manager.py
@@ -10,7 +10,7 @@ from mcdrpost.utils.exception import InvalidItem
 from mcdrpost.utils.translation import TranslationKeys
 
 if TYPE_CHECKING:
-    from mcdrpost.mcdrpost_coordinator import MCDRpostCoordinator
+    from mcdrpost.coordinator import MCDRpostCoordinator
 
 
 class PostManager: